Remove duplicated code and add utils module
